Backend/requirements.txt
appdata==2.1.2
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.3
colored==1.4.3
Django==4.1.4
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pip==21.1.2
pipreqs==0.4.11
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
setuptools==57.0.0
sqlparse==0.4.3
to-requirements.txt==1.3.0
urllib3==1.26.13
wheel==0.36.2
yarg==0.1.9
